removed bank_data.py added tests for bank.py
